Add files via upload
This file will take all the spatial predictors such as vegetation cover and climate NA variables acquired by Brendan and Sander and and get the updated CWD information from Xanthe which adjusts below ground combustion estimates.  Files are exported as a csv and  a shapefile.  New spatial predictors will be acquired with the shapefile and currently listed variables in the file will be used for model training as is without updates.